Split PropertiesPanel into Node and Edge components
